+++ src/main/java/shop/itbook/itbookshop/productgroup/product/dto/response/ProductDetailsResponseDto.java
@@ -39,7 +39,7 @@ public class ProductDetailsResponseDto {
 
     private Long fixedPrice;
 
-    private Integer discountPercent;
+    private Double discountPercent;
 
     private String fileThumbnailsUrl;
 
@@ -77,7 +77,7 @@ public class ProductDetailsResponseDto {
                                      String detailsDescription, Boolean isSelled,
                                      Boolean isForceSoldOut,
                                      Integer stock, Integer increasePointPercent, Long rawPrice,
-                                     Long fixedPrice, Integer discountPercent,
+                                     Long fixedPrice, Double discountPercent,
                                      String fileThumbnailsUrl,
                                      String isbn, Integer pageCount, LocalDateTime bookCreatedAt,
                                      Boolean isEbook, String fileEbookUrl, String publisherName,
