+++ src/main/java/shop/itbook/itbookshop/productgroup/product/dto/response/ProductSearchResponseDto.java
@@ -27,7 +27,7 @@ public class ProductSearchResponseDto {
     private String thumbnailUrl;
     private Long fixedPrice;
     private Integer increasePointPercent;
-    private Integer discountPercent;
+    private Double discountPercent;
     private Long rawPrice;
     private Long selledPrice;
 }
